CAMEL-20113 module tests/camel-itest-jms2 contained only jms component tests. They are moved to jms component and tests/camel-itest-jms2 is deleted
