+++ tests/pom.xml
@@ -40,7 +40,6 @@
         <module>test-bundles</module>
         <module>camel-itest-standalone</module>
         <module>camel-itest</module>
-        <module>camel-itest-jms2</module>
         <module>camel-partial-classpath-test</module>
         <module>camel-typeconverterscan-test</module>
     </modules>
